Move dashboard tests into right directory
